Use only native bindings
