--- .github/workflows/release.yml
@@ -1,41 +1,0 @@
-name: Build
-
-on:
-  push:
-
-jobs:
-  build-arm:
-    runs-on: macos-15
-    steps:
-      - uses: actions/checkout@v3
-      - name: Set up Node.js
-        uses: actions/setup-node@v3
-        with:
-          node-version: '18'
-      - name: Install dependencies
-        run: npm install
-      - name: Build arm artifact
-        run: npm run build:module:arm
-      - name: Upload arm artifact
-        uses: actions/upload-artifact@v3
-        with:
-          name: arm-artifact
-          path: aperture-arm.node
-
-  build-x86:
-    runs-on: macos-15-large
-    steps:
-      - uses: actions/checkout@v3
-      - name: Set up Node.js
-        uses: actions/setup-node@v3
-        with:
-          node-version: '18'
-      - name: Install dependencies
-        run: npm install
-      - name: Build x86 artifact
-        run: npm run build:module:x86
-      - name: Upload arm artifact
-        uses: actions/upload-artifact@v3
-        with:
-          name: x86-artifact
-          path: aperture-x86.node
