+++ Package.resolved
@@ -6,16 +6,7 @@
       "location" : "https://github.com/wulkano/Aperture",
       "state" : {
         "branch" : "george/rewrite-in-screen-capture-kit",
-        "revision" : "eaaee3b550c23b86e0df39ae3261d0024923ecfd"
-      }
-    },
-    {
-      "identity" : "swift-argument-parser",
-      "kind" : "remoteSourceControl",
-      "location" : "https://github.com/apple/swift-argument-parser",
-      "state" : {
-        "revision" : "41982a3656a71c768319979febd796c6fd111d5c",
-        "version" : "1.5.0"
+        "revision" : "ac1febd90238b0bbd1989beac2d7db3a4ca20f41"
       }
     },
     {
